Changed BitIOCurrentArgument to unsigned 64 bit int.
Added uint24_t and int24_t int types to the source file, to keep from
contaminating downstream projects.

Started fixing UnitTest's fake input output stuff but got stuck

Fixed an extension issue with the makefile

+++ Library/include/BufferIO.h
@@ -33,7 +33,7 @@ extern "C" {
 #pragma GCC poison bzero                                           // Not portable
 
 #pragma GCC poison strcmp                                          // misses cases that it shouldn't
-	
+
 	/*
 	typedef enum Truth {
 		YES =  true,
@@ -65,7 +65,7 @@ extern "C" {
 		BitIOGUIDSize             = BitIOUUIDSize,
     } BitIOConstants;
     
-    extern int BitIOCurrentArgument; // This HAS to start at one; Used by the Option and Input/Output parsers.
+    extern uint64_t BitIOCurrentArgument; // This HAS to start at one; Used by the Option and Input/Output parsers.
 
 	typedef enum SystemErrors {
 		SYSEmergency               = 1,
